Faking inheritance in Fedora Repository
In relational databases inheritance is handled via two tables:  table for base class and another one for the inherited class. The inherited has a foreign key to the base class and when fetching the inherited, a join is performed to get the data from base. In Fedora Repository this would mean two requests as joins are not supported. That's why fedoralink/models removes the inheritance and copies all fields from parent class to the inherited one.